blogs, footer, general redesign and organization
