+++ frontend/components/nav-bar.tsx
@@ -1,6 +1,7 @@
 import Link from "next/link"
 import Image from "next/image"
-import { CalendarButton } from "./calendar-button"
+import { Button } from "./ui/button"
+import CalendlyButton from "./calendly-button"
 
 export const NavBar = () => {
   return (
@@ -11,12 +12,16 @@ export const NavBar = () => {
         </Link>
         </div>
         <div className="flex items-center gap-4">
-          <CalendarButton variant="small" />
-          <Link
-            href="/contact"
-            className="inline-flex h-10 items-center justify-center rounded-md bg-primary/5 px-4 py-2 text-sm font-medium text-foreground transition-colors hover:bg-primary/10 focus:outline-none focus:ring-2 focus:ring-primary/20 focus:ring-offset-2 focus:ring-offset-background"
-          >
-            Contact Us
+          <Link href="/blog" className="text-sm text-muted-foreground mr-2">
+            Blog
+          </Link>
+          <CalendlyButton size="small" />
+          <Link href="/contact">
+            <Button 
+              className="bg-[#ffffff] border-2 border-[#000000] hover:bg-[#000000] hover:text-white text-black px-4 py-4 text-sm font-semibold rounded-lg w-[180px]"
+            >
+              Talk to Sales
+            </Button>
           </Link>
       </div>
     </header>
